完成 基于用户和物品的协同过滤算法 删除 不必要代码

+++ src/main/java/com/air/controller/RecommendController.java
@@ -16,7 +16,7 @@ public class RecommendController {
 
     @RequestMapping("item/{id}")
     public ResultData listRecommendByItem(@PathVariable String id) {
-        return new ResultData().success(recommendService.listRecommendByItem(Integer.valueOf(id)));
+        return new ResultData().success(recommendService.listRecommendByItem(Long.valueOf(id)));
     }
 
     @RequestMapping("user/{id}")
